Refactor the Parser and add a couple unit tests for it.
The top-level logic of the parser is (1) if it is a path, add it to the
list, (2) if it is a parameter, adjust the current values, (3) otherwise,
throw an exception. I wanted to separate the duties of the top-level
logic from the duties of the detailed scanning of each token instead
of having them all mashed together into one method. I deeply
apologize for any dudgeon this might cause.